REF: Create StorageExtensionDtype (#46537)
