+++ pandas/core/dtypes/base.py
@@ -1,7 +1,6 @@
 """
 Extend pandas with custom array types.
 """
-
 from __future__ import annotations
 
 from typing import (
@@ -14,6 +13,7 @@ from typing import (
 
 import numpy as np
 
+from pandas._libs import missing as libmissing
 from pandas._libs.hashtable import object_hash
 from pandas._typing import (
     DtypeObj,
@@ -389,6 +389,32 @@ class ExtensionDtype:
         Can arrays of this dtype hold NA values?
         """
         return True
+
+
+class StorageExtensionDtype(ExtensionDtype):
+    """ExtensionDtype that may be backed by more than one implementation."""
+
+    name: str
+    na_value = libmissing.NA
+    _metadata = ("storage",)
+
+    def __init__(self, storage=None) -> None:
+        self.storage = storage
+
+    def __repr__(self):
+        return f"{self.name}[{self.storage}]"
+
+    def __str__(self):
+        return self.name
+
+    def __eq__(self, other: Any) -> bool:
+        if isinstance(other, self.type) and other == self.name:
+            return True
+        return super().__eq__(other)
+
+    def __hash__(self) -> int:
+        # custom __eq__ so have to override __hash__
+        return super().__hash__()
 
 
 def register_extension_dtype(cls: type_t[ExtensionDtypeT]) -> type_t[ExtensionDtypeT]:
